+++ src/client/sprites/ImageSprite.ts
@@ -1,4 +1,4 @@
-import { canvas, ctx, penCtx } from '@main/canvas.ts';
+import { ctx, penCtx } from '@main/canvas.ts';
 import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
 import TSCMath from '@main/TSCMath.ts';
 
@@ -63,11 +63,7 @@ export default class ImageSprite extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
-        c.translate(cX, cY);
-        c.rotate(this.toRadians(this.dir));
-        c.translate(-this.pivot[0] * this.size, this.pivot[1] * this.size);
+        this.applyDrawTransform(c);
 
         c.strokeStyle = this.outlineColor;
         c.lineWidth = this.outlineWidth;