+++ src/client/sprites/Joystick.ts
@@ -48,11 +48,7 @@ export default class Joystick extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
-        c.translate(cX, cY);
-        c.rotate(this.toRadians(this.dir));
-        c.translate(-this.pivot[0] * this.size, this.pivot[1] * this.size);
+        this.applyDrawTransform(c);
 
         // Base
         const path = this.getCachedPath();