+++ src/client/sprites/Rectangle.ts
@@ -1,5 +1,5 @@
 import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
-import { ctx, canvas, penCtx } from '@main/canvas.ts';
+import { ctx, penCtx } from '@main/canvas.ts';
 import TSCMath from '@main/TSCMath.ts';
 
 export interface RectangleOptions extends SpriteOptions {
@@ -58,11 +58,7 @@ export default class Rectangle extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
-        c.translate(cX, cY);
-        c.rotate(this.toRadians(this.dir));
-        c.translate(-this.pivot[0] * this.size, this.pivot[1] * this.size);
+        this.applyDrawTransform(c);
 
         const path = this.getCachedPath();
 
@@ -111,7 +107,7 @@ export default class Rectangle extends Sprite {
     constructor(options?: RectangleOptions) {
         super(options);
 
-        this.width = options?.width ?? 50;
+        this.width = options?.width ?? 75;
         this.height = options?.height ?? 50;
         this.color = options?.color ?? 'black';
         this.outlineColor = options?.outlineColor ?? 'black';