+++ src/client/sprites/RegularPolygon.ts
@@ -1,6 +1,5 @@
-import { canvas, ctx, penCtx } from '@main/canvas.ts';
+import { ctx, penCtx } from '@main/canvas.ts';
 import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
-import TSCMath from '@main/TSCMath.ts';
 
 export interface RegularPolygonOptions extends SpriteOptions {
     sides?: number;
@@ -53,11 +52,7 @@ export default class RegularPolygon extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
-        c.translate(cX, cY);
-        c.rotate(TSCMath.toRadians(this.dir));
-        c.translate(-this.pivot[0] * this.size, this.pivot[1] * this.size);
+        this.applyDrawTransform(c);
 
         const path = this.getCachedPath();